Parse redirect and subsequent data packet together
Since the data packet following a redirect may not be in the buffer at
the time of parsing the redirect packet itself, this is handled as a
special case in the packet decoder.

When a redirect packet is decoded, it reads from the socket until it has
the following data packet which contains the redirect connection
information.